requirements.txt
pytest==7.1.1
selenium==4.1.5
allure-pytest==2.9.45
pytest-xdist==2.5.0